pyproject.toml
[tool.poetry]
name = "semgr8s"
version = "0.0.0"
description = "Semgrep-based Policy controller for Kubernetes."
authors = ["Christoph Hamsen <christoph.hamsen@securesystems.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
APScheduler = "3.10.4"
Flask = "3.0.2"
gunicorn = "21.2.0"
PyYAML = "6.0.1"
semgrep = "1.65.0"
jsonpatch = "1.33"

[tool.poetry.group.docs]
optional = false

[tool.poetry.group.docs.dependencies]
mkdocs-material = "9.5.13"
mike = "2.0.0"

[tool.poetry.group.dev]
optional = false

[tool.poetry.group.dev.dependencies]
bandit = "1.7.8"
bandit-sarif-formatter = "1.1.1"
black = "24.2.0"
pylint = "3.1.0"

[tool.poetry.group.test]
optional = false

[tool.poetry.group.test.dependencies]
pytest = "8.1.1"
pytest-cov = "4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
